Creating the React project
There is an error parsing the web configuration file at line 1. A document must contain exactly once root element when try to update the target to .net framework 4.8.
Fixing by: adding the following AppContext switch to the <runtime> section of the app config file and setting it to false (edit the Web.config file)